recycle the genome fa.gz file

+++ BLAST_utils.py
@@ -7,6 +7,20 @@ import gzip
 from subprocess import Popen
 from Bio import SeqIO
 
+def check_gzip_integrity(filepath):
+    import gzip
+    import os
+    if not os.path.isfile(filepath):
+        return False #file doesn't exist
+    chunksize = 1024 * 1024
+    with gzip.open(filepath) as g:
+        try:
+            while g.read(chunksize):
+                pass
+            return True
+        except:
+            return False #file corrupted
+
 def check_blastDB_human(ref):
     #set BLAST binary dir
     BLAST_bin = ""
@@ -42,7 +56,7 @@ def check_blastDB_human(ref):
     BLAST_db_path = os.path.join(BLASTDB_DIR, BLAST_db)
     
     fa_gz = fa + ".gz"
-    fa_gz_path = os.path.join(BLASTDB_DIR, fa_gz)
+    fa_gz_path = os.path.join(BLASTDB_DIR, ref, fa_gz)
     fa = fa_gz.rstrip(".gz")
 
     url = prefix + fa_gz
@@ -53,29 +67,32 @@ def check_blastDB_human(ref):
            os.path.isfile(f"{BLAST_db_path}.not") and os.path.isfile(f"{BLAST_db_path}.ntf") and
            os.path.isfile(f"{BLAST_db_path}.nto")):
 
-        print(f"BLAST database not found\nDownloading {url}\nThis is a one-time process (this may take several minutes, if your internet connection is spotty, the download may fail and display an error mentioning an unexpected end of the stream)", flush=True)
+        print(f"BLAST database not found, automatically generating BLAST databases (this is a one-time process )", flush=True)
 
-        #download gzip file
-        if os.path.isfile(fa_gz_path): #remove gzip file, b/c it could be a partial file
-            os.remove(fa_gz_path)
-        # Download the file from `url` and save it locally under `file_name`:
-        with urllib.request.urlopen(url) as response, open(fa_gz_path, 'wb') as out_file:
-            shutil.copyfileobj(response, out_file)
+        # check if the gz file exists, and also check integrity
+        # download gzip file, download the file from `url` and save it locally under `file_name`:
+        if not check_gzip_integrity(fa_gz_path): #check gzip file integrity
+            print(f"...(re)downloading {fa_gz_path} sequence file, (this may take several minutes, if your internet connection is spotty, the download may fail and display an error mentioning an unexpected end of the stream)", flush=True)
+            if os.path.isfile(fa_gz_path): #remove gzip file, b/c it could be a partial file
+                os.remove(fa_gz_path)
+            with urllib.request.urlopen(url) as response, open(fa_gz_path, 'wb') as out_file: #download gzip file
+                shutil.copyfileobj(response, out_file)  
 
         #unzip file
         if os.path.isfile(BLAST_db_path): #remove unzipped file, b/c it could be a partial file
             os.remove(BLAST_db_path)
-        print(f"unzipping {fa_gz_path}, this is a one-time process", flush=True)
+        print(f"...Unzipping {fa_gz_path}", flush=True)
         with gzip.open(fa_gz_path) as f_in:
             with open(BLAST_db_path, "wb") as f_out:
                 shutil.copyfileobj(f_in, f_out)
 
         #make blastdb
         cmd = [f"{BLAST_bin}makeblastdb{exe_suffix}", "-dbtype", "nucl", "-in", f"{BLAST_db_path}"]
-        print(f"Making BLAST database, This is a one-time process", flush=True)
+        print(f"...Making BLAST database, this is a one-time process", flush=True)
         p = Popen(cmd, cwd = os.getcwd(), universal_newlines=True)
         p.communicate()  # now wait plus that you can send commands to process
 
-        os.remove(fa_gz_path)
+        os.remove(fa_gz_path) # save the gz file for making BLAST database
         os.remove(BLAST_db_path)
+        print(f"Finished generating BLAST databases, this is a one-time process)", flush=True)
     return([BLAST_bin, exe_suffix, BLAST_db_path])
